+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Bump CUDA base image version to 12.3.2
 - Add `cdi-cri` device list strategy. This uses the CDIDevices CRI field to request CDI devices instead of annotations.
+- Set MPS memory limit by device index and not device UUID. This is a workaround for an issue where
+  these limits are not applied for devices if set by UUID.
 
 ### Version v0.15.0-rc.1
 - Import GPU Feature Discovery into the GPU Device Plugin repo. This means that